moved the driver into the nvidia file and updated docs
