+++ README.md
@@ -101,7 +101,7 @@ The following devices are supported by this version of liquidctl.  See each guid
 |:--- |:---:|:--- |
 | Gigabyte RGB Fusion 2.0 Motherboards | [guide](docs/gigabyte-rgb-fusion2-guide.md) | <sup>_E_</sup> |
 | EVGA GTX 1080 FTW | [guide](docs/nvidia-guide.md) | <sup>_EUXN_</sup> |
-| ASUS Strix RTX 2080 Ti OC | [guide](docs/asus-rog-guide.md) | <sup>_EUXN_</SUP> |
+| ASUS Strix RTX 2080 Ti OC | [guide](docs/anvidia-guide.md) | <sup>_EUXN_</SUP> |
 | NZXT HUE 2, HUE 2 Ambient | [guide](docs/nzxt-hue2-guide.md) | |
 
 <sup>_L_</sup> _Requires the `--legacy-690lc` flag._
