+++ docs/nvidia-guide.md
@@ -16,6 +16,7 @@ Jump to a specific card:
 
 * _Series 10/Pascal:_
     - [EVGA GTX 1080 FTW](#evga-gtx-1080-ftw)
+    - [ASUS Strix RTX 2080 Ti OC](#asus-strix-rtx-2080-ti-oc)
 <!-- - [EVGA GTX 1070 FTW](#evga-gtx-1070-1080-ftw) -->
 <!-- - [EVGA GTX 1080 FTW](#evga-gtx-1070-1080-ftw) -->
 
@@ -76,3 +77,73 @@ preferable, if the use case allows for them.
 ```
 $ liquidctl set led color fixed 00ff00 --non-volatile --unsafe=smbus,evga_pascal
 ```
+
+
+## ASUS Strix RTX 2080 Ti OC
+
+Experimental. Only RGB lighting supported.
+
+Unsafe features:
+
+- `smbus`: enable SMBus support; SMBus devices may not tolerate writes or reads they do not expect
+- `rog_turing`: enable access to the specific graphics cars
+
+
+### Initialization
+
+Not required for this device.
+
+### Retrieving the current color mode and LED color
+
+In verbose mode `status` reports the current RGB lighting settings.
+
+```
+# liquidctl status -v --unsafe=smbus,rog_turing
+
+ASUS Strix RTX 2080 Ti OC (experimental)
+├── Mode      Fixed
+└── Color    ff0000
+```
+
+## Controlling the LED
+
+This GPU only has one led that can be set.
+
+
+The table bellow summarizes the available channels, modes, and their associated
+maximum number of colors for each device family.
+
+| Channel    | Mode          | colors  |
+| ---------- | ------------- | ------- |
+| `led`      | `off`         | 0       |
+| `led`      | `fixed`       | 1       |
+| `led`      | `flash`       | 1       |
+| `led`      | `breathing`   | 1       |
+| `led`      | `rainbow`     | 0       |
+
+The `off` mode is simply an alias for `fixed 000000`.
+
+```
+# liquidctl --unsafe=smbus,rog_turing set led color off
+# liquidctl --unsafe=smbus,rog_turing set led color fixed ff8000
+# liquidctl --unsafe=smbus,rog_turing set led color fixed "hsv(90,85,70)"
+# liquidctl --unsafe=smbus,rog_turing set led color fixed <1  color>
+                                          ^^^       ^^^^^   ^^^
+                                        channel      mode   color
+```
+
+Each color can be specified using any of the [supported formats](../README.md#supported-color-specification-formats).
+
+
+The settings configured on the device are normally volatile, and are
+cleared whenever the graphics card looses power (ie. unplugged, not power off).
+
+It is possible to store them in non-volatile controller memory by
+passing `--non-volatile`.  But as this memory has some unknown yet
+limited maximum number of write cycles, volatile settings are
+preferable, if the use case allows for them.
+
+```
+$ liquidctl set led color fixed 00ff00 --non-volatile --unsafe=smbus,rog_turing
+```
+
